no changes, added some comments to help me out.
i will be changing the buffer_ptr data structure shortly to enable SFF
scheduling, right now we have a forced FIFO scheduling policy.

+++ trunk/p5/request.c
@@ -28,7 +28,7 @@ void requestError(int fd, char *cause, char *errnum, char *shortmsg, char *longm
     Rio_writen(fd, buf, strlen(buf));
     printf("%s", buf);
     
-    sprintf(buf, "Content-Length: %d\r\n\r\n", strlen(body));
+    sprintf(buf, "Content-Length: %d\r\n\r\n", (int) strlen(body));
     Rio_writen(fd, buf, strlen(buf));
     printf("%s", buf);
     
@@ -48,7 +48,7 @@ void requestReadhdrs(rio_t *rp)
     
     Rio_readlineb(rp, buf, MAXLINE);
     while (strcmp(buf, "\r\n")) {
-	Rio_readlineb(rp, buf, MAXLINE);
+        Rio_readlineb(rp, buf, MAXLINE);
     }
     return;
 }
@@ -62,24 +62,24 @@ int requestParseURI(char *uri, char *filename, char *cgiargs)
     char *ptr;
     
     if (!strstr(uri, "cgi")) {
-	// static
-	strcpy(cgiargs, "");
-	sprintf(filename, ".%s", uri);
-	if (uri[strlen(uri)-1] == '/') {
-	    strcat(filename, "home.html");
-	}
-	return 1;
+        // static
+        strcpy(cgiargs, "");
+        sprintf(filename, ".%s", uri);
+        if (uri[strlen(uri)-1] == '/') {
+            strcat(filename, "home.html");
+        }
+        return 1;
     } else {
-	// dynamic
-	ptr = index(uri, '?');
-	if (ptr) {
-	    strcpy(cgiargs, ptr+1);
-	    *ptr = '\0';
-	} else {
-	    strcpy(cgiargs, "");
-	}
-	sprintf(filename, ".%s", uri);
-	return 0;
+        // dynamic
+        ptr = index(uri, '?');
+        if (ptr) {
+            strcpy(cgiargs, ptr+1);
+            *ptr = '\0';
+        } else {
+            strcpy(cgiargs, "");
+        }
+        sprintf(filename, ".%s", uri);
+        return 0;
     }
 }
 
@@ -89,18 +89,18 @@ int requestParseURI(char *uri, char *filename, char *cgiargs)
 void requestGetFiletype(char *filename, char *filetype)
 {
     if (strstr(filename, ".html")) 
-	strcpy(filetype, "text/html");
+        strcpy(filetype, "text/html");
     else if (strstr(filename, ".gif")) 
-	strcpy(filetype, "image/gif");
+        strcpy(filetype, "image/gif");
     else if (strstr(filename, ".jpg")) 
-	strcpy(filetype, "image/jpeg");
+        strcpy(filetype, "image/jpeg");
     else 
-	strcpy(filetype, "test/plain");
+        strcpy(filetype, "test/plain");
 }
 
 void requestServeDynamic(request_type request, char *filename, char *cgiargs, thread_info_type* thread_info)
 {
-	int fd = request.conn_fd;
+        int fd = request.conn_fd;
     char buf[MAXLINE], *emptylist[] = {NULL};
     
     // The server does only a little bit of the header.  
@@ -125,11 +125,11 @@ void requestServeDynamic(request_type request, char *filename, char *cgiargs, th
     Rio_writen(fd, buf, strlen(buf));
     
     if (Fork() == 0) {
-	/* Child process */
-	Setenv("QUERY_STRING", cgiargs, 1);
-	/* When the CGI process writes to stdout, it will instead go to the socket */
-	Dup2(fd, STDOUT_FILENO);
-	Execve(filename, emptylist, environ);
+        /* Child process */
+        Setenv("QUERY_STRING", cgiargs, 1);
+        /* When the CGI process writes to stdout, it will instead go to the socket */
+        Dup2(fd, STDOUT_FILENO);
+        Execve(filename, emptylist, environ);
     }
     Wait(NULL);
 }
@@ -137,7 +137,7 @@ void requestServeDynamic(request_type request, char *filename, char *cgiargs, th
 
 void requestServeStatic(request_type request, char *filename, int filesize, thread_info_type* thread_info) 
 {
-	int fd = request.conn_fd;
+        int fd = request.conn_fd;
     int srcfd;
     char *srcp, filetype[MAXLINE], buf[MAXBUF];
     char tmp = 0;
@@ -161,7 +161,7 @@ void requestServeStatic(request_type request, char *filename, int filesize, thre
     // When you time this, you will see that the first time a client 
     //requests a file, the read is much slower than subsequent requests.
     for (i = 0; i < filesize; i++) {
-	tmp += *(srcp + i);
+        tmp += *(srcp + i);
     }
     double read_end_time = GetTime();
     request.Stat_req_read = read_end_time - read_start_time;
@@ -179,6 +179,8 @@ void requestServeStatic(request_type request, char *filename, int filesize, thre
     sprintf(buf, "%s Stat-req-complete: %f\r\n", buf, request.Stat_req_complete);
     sprintf(buf, "%s Stat-thread-id: %d\r\n", buf, thread_info -> Stat_thread_id);
     sleep(1);
+    
+    // TODO what is this line for?
     for(i = 0; i < 1e7; i++){
     }
     sprintf(buf,"%s yangsuli debug", buf);
@@ -194,18 +196,18 @@ void requestServeStatic(request_type request, char *filename, int filesize, thre
     //  Writes out to the client socket the memory-mapped file 
     Rio_writen(fd, srcp, filesize);
     Munmap(srcp, filesize);
-    
+
 }
 
 // handle a request
 void requestHandle(request_type request, thread_info_type* thread_info)
 {
 
-	thread_info -> Stat_thread_count ++;
+        thread_info -> Stat_thread_count ++;
 
     int fd = request.conn_fd;    
     int is_static;
-    struct stat sbuf;
+    struct stat sbuf;  // see man 2 stat: off_t sbuf.st_size /* file size, in bytes */
     char buf[MAXLINE], method[MAXLINE], uri[MAXLINE], version[MAXLINE];
     char filename[MAXLINE], cgiargs[MAXLINE];
     rio_t rio;
@@ -217,32 +219,32 @@ void requestHandle(request_type request, thread_info_type* thread_info)
     printf("%s %s %s\n", method, uri, version);
     
     if (strcasecmp(method, "GET")) {
-	requestError(fd, method, "501", "Not Implemented", 
-		     "CS537 Server does not implement this method");
-	return;
+        requestError(fd, method, "501", "Not Implemented", 
+                     "CS537 Server does not implement this method");
+        return;
     }
     requestReadhdrs(&rio);
     
     is_static = requestParseURI(uri, filename, cgiargs);
     if (stat(filename, &sbuf) < 0) {
-	requestError(fd, filename, "404", "Not found", "CS537 Server could not find this file");
-	return;
+        requestError(fd, filename, "404", "Not found", "CS537 Server could not find this file");
+        return;
     }
     
     if (is_static) {
-	    thread_info -> Stat_thread_static ++;
-	if (!(S_ISREG(sbuf.st_mode)) || !(S_IRUSR & sbuf.st_mode)) {
-	    requestError(fd, filename, "403", "Forbidden", "CS537 Server could not read this file");
-	    return;
-	}
-	requestServeStatic(request, filename, sbuf.st_size, thread_info);
+            thread_info -> Stat_thread_static ++;
+        if (!(S_ISREG(sbuf.st_mode)) || !(S_IRUSR & sbuf.st_mode)) {
+            requestError(fd, filename, "403", "Forbidden", "CS537 Server could not read this file");
+            return;
+        }
+        requestServeStatic(request, filename, sbuf.st_size, thread_info);
     } else {
-	    thread_info -> Stat_thread_dynamic ++;
-	if (!(S_ISREG(sbuf.st_mode)) || !(S_IXUSR & sbuf.st_mode)) {
-	    requestError(fd, filename, "403", "Forbidden", "CS537 Server could not run this CGI program");
-	    return;
-	}
-	requestServeDynamic(request, filename, cgiargs, thread_info);
+            thread_info -> Stat_thread_dynamic ++;
+        if (!(S_ISREG(sbuf.st_mode)) || !(S_IXUSR & sbuf.st_mode)) {
+            requestError(fd, filename, "403", "Forbidden", "CS537 Server could not run this CGI program");
+            return;
+        }
+        requestServeDynamic(request, filename, cgiargs, thread_info);
     }
 }
